Setting the password only makes sense for the first node
The ALTER KEYSPACE query was executed unconditionally
on every startup, now it is only executed only during
the first startup.

When joining a cluster, there is already a seed node
which should have a non-default password.

The number of replicas for the KS system_auth should
be adjusted, when you create a multi-node cluster.